[MOD] Доделал задание.

+++ JavaLessons/Part_3/FileFinder/src/main/java/ru/kostikov/filefinder/FileMain.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.*;
+import java.util.Optional;
 
 
 /**
@@ -14,21 +15,25 @@ public class FileMain extends SimpleFileVisitor<Path> {
     public static void main(String[] args) {
         Ui argsParser = new Ui();
 
-        ParseValue parseValue = argsParser.argsParse(args);
-        File out = new File(parseValue.fileOutName);
-        FileOutputStream output = null;
+        Optional<ParseValue> parseValue = argsParser.argsParse(args);
 
-        if(out.exists()){
-            out.delete();
+        if(parseValue.isPresent()){
+            File out = new File(parseValue.get().fileOutName);
+            FileOutputStream output = null;
+
+            if(out.exists()){
+                out.delete();
+            }
+            try {
+                out.createNewFile();
+                output = new FileOutputStream(out);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            FileFinder fileFinder = new FileFinder(parseValue.get().matchTemplate, output);
+            fileFinder.findFiles(Paths.get(parseValue.get().url));
         }
-        try {
-            out.createNewFile();
-            output = new FileOutputStream(out);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        FileFinder fileFinder = new FileFinder(parseValue.matchTemplate, output);
-        fileFinder.findFiles(Paths.get(parseValue.url));
+
     }
 }
